+++ src/main/redux/sagas/i18n.ts
@@ -9,7 +9,7 @@ import * as debug_ from "debug";
 import { LocaleConfigIdentifier, LocaleConfigValueType } from "readium-desktop/common/config";
 import { error } from "readium-desktop/common/error";
 import { i18nActions } from "readium-desktop/common/redux/actions";
-import { callTyped } from "readium-desktop/common/redux/typed-saga";
+import { callTyped } from "readium-desktop/common/redux/sagas/typed-saga";
 import { ConfigRepository } from "readium-desktop/main/db/repository/config";
 import { diMainGet } from "readium-desktop/main/di";
 import { all, call, takeLeading } from "redux-saga/effects";
